added some more of the icon ui

+++ Libus/App/views.py
@@ -177,7 +177,19 @@ def posts(request):
     
     return paginator.get_paginated_response(serializer.data)
 
+@api_view(['Get'])
+def HasALike(request, id):
+    if(request.user.is_authenticated == False):
+        return redirect("/login")
+    post = Post.objects.get(id=id)
+    if request.user in post.liked.all():
+        return HttpResponse(True)
+    return HttpResponse(False)
+
+@api_view(['Get'])
 def delete(request, id):
+    if(request.user.is_authenticated == False):
+        return redirect("/login")
     post = Post.objects.get(id=id)
     if post.author.username == request.user.username:
         post.delete()
@@ -192,3 +204,16 @@ def get_users(request, username):
         users = users[:99]
     user_list = [{"username": user.username} for user in users]
     return HttpResponse(json.dumps(user_list), content_type="application/json")
+
+@api_view(['Get'])
+def like_or_dislike(request, id):
+    if(request.user.is_authenticated == False):
+        return redirect("/login")
+    post = Post.objects.get(id=id)
+    if request.user in post.liked.all():
+        post.liked.remove(request.user)
+    else:
+        post.liked.add(request.user)
+    if post.author.username == request.user.username:
+        post.delete()
+    return HttpResponse("")
